# docs/swagger.yaml
basePath: /api/v1/
definitions:
...
  controller.CommonResult:
    properties:
      code:
        type: integer
      data:
        type: object
      message:
        type: string
    type: object
  controller.Page:
    properties:
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  controller.ResultWithPage:
    properties:
      code:
        type: integer
      data:
        type: object
      message:
        type: string
      page:
        $ref: '#/definitions/controller.Page'
    type: object
...
  message.CheckClusterRsp:
    properties:
      checkId:
        type: string
      workFlowId:
        description: Asynchronous task workflow ID
        type: string
    type: object
...
  message.CheckPlatformReq:
    properties:
      displayMode:
        type: string
    type: object
  message.CheckPlatformRsp:
    properties:
      checkId:
        type: string
      workFlowId:
        description: Asynchronous task workflow ID
        type: string
    type: object
...
  message.CreateUserReq:
    properties:
      email:
        type: string
      name:
        type: string
      nickname:
        type: string
      password:
        type: string
      phone:
        type: string
      tenantId:
        type: string
    required:
    - email
    - name
    - password
    type: object
  message.CreateUserResp:
    type: object
...
  message.DeleteUserReq:
    type: object
  message.DeleteUserResp:
    type: object
...
  message.GetCheckReportRsp:
    properties:
      reportInfo:
        type: object
    type: object
...
  message.GetUserResp:
    properties:
      user:
        $ref: '#/definitions/structs.UserInfo'
    type: object
...
  message.QueryCheckReportsRsp:
    properties:
      reportMetas:
        additionalProperties:
          $ref: '#/definitions/structs.CheckReportMeta'
        type: object
    type: object
...
  message.QueryUserReq:
    properties:
      page:
        description: Current page location
        type: integer
      pageSize:
        description: Number of this request
        type: integer
    type: object
  message.QueryUserResp:
    properties:
      users:
        additionalProperties:
          $ref: '#/definitions/structs.UserInfo'
        type: object
    type: object
...
  message.UpdateUserPasswordReq:
    properties:
      id:
        type: string
      password:
        type: string
    required:
    - password
    type: object
  message.UpdateUserPasswordResp:
    type: object
  message.UpdateUserProfileReq:
    properties:
      email:
        type: string
      nickname:
        type: string
      phone:
        type: string
    type: object
  message.UpdateUserProfileResp:
    type: object
...
  structs.CheckReportMeta:
    properties:
      checkID:
        type: string
      createAt:
        type: string
      creator:
        type: string
      status:
        type: string
      type:
        type: string
    type: object
...
  structs.UserInfo:
    properties:
      createAt:
        type: string
      creator:
        type: string
      defaultTenantId:
        type: string
      email:
        type: string
      id:
        type: string
      names:
        items:
          type: string
        type: array
      nickname:
        type: string
      phone:
        type: string
      status:
        type: string
      tenantIds:
        items:
          type: string
        type: array
      updateAt:
        type: string
    type: object
...
host: localhost:4116
info:
  contact:
    email: zhangpeijin@pingcap.com
    name: zhangpeijin
  description: EM UI API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: EM UI API
  version: "1.0"
paths:
...
  /platform/check:
    post:
      consumes:
      - application/json
      description: platform check
      parameters:
      - description: check platform
        in: body
        name: checkPlatformReq
        schema:
          $ref: '#/definitions/message.CheckPlatformReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controller.ResultWithPage'
            - properties:
                data:
                  $ref: '#/definitions/message.CheckPlatformRsp'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.CommonResult'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonResult'
      security:
      - ApiKeyAuth: []
      summary: platform check
      tags:
      - platform
  /platform/check/{clusterId}:
    post:
      consumes:
      - application/json
      description: platform check cluster
      parameters:
      - description: cluster id
        in: path
        name: clusterId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controller.ResultWithPage'
            - properties:
                data:
                  $ref: '#/definitions/message.CheckClusterRsp'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.CommonResult'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonResult'
      security:
      - ApiKeyAuth: []
      summary: platform check cluster
      tags:
      - platform
...
  /platform/report/{checkId}:
    get:
      consumes:
      - application/json
      description: get check report based on checkId
      parameters:
      - description: check id
        in: path
        name: checkId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controller.ResultWithPage'
            - properties:
                data:
                  $ref: '#/definitions/message.GetCheckReportRsp'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.CommonResult'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonResult'
      security:
      - ApiKeyAuth: []
      summary: get check report based on checkId
      tags:
      - platform
  /platform/reports:
    get:
      consumes:
      - application/json
      description: get query check reports
      parameters:
      - description: Current page location
        in: query
        name: page
        type: integer
      - description: Number of this request
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controller.ResultWithPage'
            - properties:
                data:
                  $ref: '#/definitions/message.QueryCheckReportsRsp'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.CommonResult'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonResult'
      security:
      - ApiKeyAuth: []
      summary: query check reports
      tags:
      - platform
...
  /users/:
    get:
      consumes:
      - application/json
      description: query all user profile
      parameters:
      - description: query user profile request parameter
        in: body
        name: queryUserRequest
        required: true
        schema:
          $ref: '#/definitions/message.QueryUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controller.ResultWithPage'
            - properties:
                data:
                  $ref: '#/definitions/message.QueryUserResp'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.CommonResult'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonResult'
      security:
      - ApiKeyAuth: []
      summary: queries all user profile
      tags:
      - platform
    post:
      consumes:
      - application/json
      description: created user
      parameters:
      - description: create user request parameter
        in: body
        name: createUserReq
        required: true
        schema:
          $ref: '#/definitions/message.CreateUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controller.CommonResult'
            - properties:
                data:
                  $ref: '#/definitions/message.CreateUserResp'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.CommonResult'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonResult'
      security:
      - ApiKeyAuth: []
      summary: created  user
      tags:
      - platform
  /users/{userId}:
    delete:
      consumes:
      - application/json
      description: delete user
      parameters:
      - description: user id
        in: path
        name: userId
        required: true
        type: string
      - description: delete user request parameter
        in: body
        name: deleteUserReq
        required: true
        schema:
          $ref: '#/definitions/message.DeleteUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controller.CommonResult'
            - properties:
                data:
                  $ref: '#/definitions/message.DeleteUserResp'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.CommonResult'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonResult'
      security:
      - ApiKeyAuth: []
      summary: delete user
      tags:
      - platform
    get:
      consumes:
      - application/json
      description: get user profile
      parameters:
      - description: user id
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controller.CommonResult'
            - properties:
                data:
                  $ref: '#/definitions/message.GetUserResp'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.CommonResult'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonResult'
      security:
      - ApiKeyAuth: []
      summary: get user profile
      tags:
      - platform
  /users/{userId}/password:
    post:
      consumes:
      - application/json
      description: update user password
      parameters:
      - description: user id
        in: path
        name: userId
        required: true
        type: string
      - description: query user password request parameter
        in: body
        name: UpdateUserPasswordRequest
        required: true
        schema:
          $ref: '#/definitions/message.UpdateUserPasswordReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controller.CommonResult'
            - properties:
                data:
                  $ref: '#/definitions/message.UpdateUserPasswordResp'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.CommonResult'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonResult'
      security:
      - ApiKeyAuth: []
      summary: update user password
      tags:
      - platform
  /users/{userId}/update_profile:
    post:
      consumes:
      - application/json
      description: update user profile
      parameters:
      - description: user id
        in: path
        name: userId
        required: true
        type: string
      - description: query user profile request parameter
        in: body
        name: updateUserProfileRequest
        required: true
        schema:
          $ref: '#/definitions/message.UpdateUserProfileReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controller.CommonResult'
            - properties:
                data:
                  $ref: '#/definitions/message.UpdateUserProfileResp'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.CommonResult'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonResult'
      security:
      - ApiKeyAuth: []
      summary: update user profile
      tags:
      - platform
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"